+++ Tests/SwiftDockerLibTests/DockerCommandTests.swift
@@ -5,6 +5,24 @@ enum TestError: Error {
   case failed
 }
 
+struct MockDockerCommand: DockerCommand {
+  var options: CLIOptions
+  var shell: ShellProtocol.Type = MockShell.self
+  var output: OutputDestination = MockOutput()
+}
+
+class DockerCommandProtocolTests: XCTestCase {
+  func testDockerRunCommandsVapor() {
+    let expectedCmd = """
+    docker run --rm --mount type=bind,source=/tmp/my-package,target=/package --mount type=volume,source=swiftdockercli-my-package,target=/package/.build --workdir /package  swift:latest swift run --enable-test-discovery Run serve --hostname 0.0.0.0 --env production
+    """
+    let options = CLIOptions(swift: "latest", path: "/tmp/my-package", verbose: false, clean: false, args: [], seedBuildFolder: false)
+    let cmd = MockDockerCommand(options: options)
+    let dockerCmd = cmd.makeDockerRunCommand(cmd: "swift run --enable-test-discovery Run serve --hostname 0.0.0.0 --env production", labels: "")
+    XCTAssertEqual(dockerCmd, expectedCmd)
+  }
+}
+
 class DockerImageTests: XCTestCase {
   func testInitailizedWithVersion() throws {
     let version = "4.1"