+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `swift docker test` defaults to using the swift-tools-version specified in Package.swift
 
 ### Added
+- `swift docker vapor` - Run your vapor application in a container
 - `swift docker run` - mirrors swift run command
 - `swift docker build` - mirrors swift build
 - `swift docker build-image` builds a docker image of your package